Work in progress: Smaller thread ids

+++ javassist-inst/qtrtool-java13/src/main/java/veroy/research/qtrtool/javassist/QTRProxy.java
@@ -449,7 +449,7 @@ public class QTRProxy {
         }
     }
 
-    protected static Long getThreadId(Long objHashCode) {
+    protected static int getThreadId(int objHashCode) {
         if (threadIdMap.containsKey(objHashCode)) {
             return threadIdMap.get(objHashCode);
         } else {
